Remove duplicated test code
