+++ csunplugged/tests/config/templatetags/test_translate_url.py
@@ -15,7 +15,7 @@ AVAILABLE_LANGUAGES_EN = (
     ("en", "English"),
 )
 
-VALID_PATH = "/chapters/chapter-name/"
+VALID_PATH = "/topics/topic-name/"
 INVALID_PATH = "/this/is/an/invalid/url"
