--- csunplugged/tests/general/templatetags/__init__.py
@@ -1,1 +1,0 @@
-"""Module for tests of the template tags in the general application."""
